Use setuptools entry_points console_scripts to install command line tools
This provides a more consistent experience across platforms. It allows the ".py" extension to be omitted from the command line tools, and setuptools ensures the correct python interpreter is called and makes sure everything works on Windows as well.

The scripts are now just automatically generated by referring to a function in the package, so dumppdf.py, pdf2txt.py and latin2ascii.py have just been moved into the main pdfminer package. It might be worth considering something like a "cli" sub-package, especially if the number of command line tools grows.