Fix XML output to include spaces in "text" attribute
Sometimes whitespace is not explicitly defined in a PDF - it is just a gap between two characters. PDFMiner detects this and automatically inserts virtual spaces as "LTAnon" objects. Previously these weren't included in the "text" attribute of "textline" elements, which resulted in output that was missing all whitespace for some PDFs. This fix adds a single space for each "LTAnon" element, regardless of whether it contains multiple spaces or other whitespace characters. Leading and trailing spaces are also stripped off.